fix(product): repair shell-actions imports and split the agents-pane facade under the size threshold
Two more stale paths from the agent-ops stack transplant (#1773): the
footer controls imported useWorkspaceShellActions from the pre-#1851
provider module, which no longer exports it — web and desktop typecheck
builds fail on main. Point them at the hook module every other caller
uses.

use-agents-pane.ts also landed at 421 lines, tripping the strict
LARGE_FRONTEND_FILE structure gate on main. Split the three roster
queries (and their staleness-backstop poll policy) into
use-agents-pane-roster-queries.ts along the data-fetching seam.
