+++ apps/packages/product-client/src/components/workspace/chat/input/WorkspaceOpenInWebFooterControl.tsx
@@ -1,5 +1,5 @@
 import { ExternalLink } from "#product/primitives/icons/core";
-import { useWorkspaceShellActions } from "#product/components/workspace/shell/providers/WorkspaceShellActionsContext";
+import { useWorkspaceShellActions } from "#product/hooks/workspaces/workflows/use-workspace-shell-actions";
 import { ComposerControlButton } from "#product/primitives/patterns/composer/ComposerControlButton";
 
 export function WorkspaceOpenInWebFooterControl() {
